Move CalculatePanelAreas() from utils/display.cpp to control.cpp
