+++ Source/control.h
@@ -72,6 +72,7 @@ void ToggleSpell(int slot);
 void AddPanelString(const char *str);
 void ClearPanel();
 void DrawPanelBox(const Surface &out, SDL_Rect srcRect, Point targetPosition);
+void CalculatePanelAreas(void);
 bool IsPanelOpen(UiPanels panel);
 Point GetPanelPosition(UiPanels panel, Point offset = { 0, 0 });
